cdi spec: move cdi spec definition to a dedicated config.rs
As original mod.rs where cdi spec defined, I think we should
move it to a dedicated config.rs for CDI spec which ensures
adherence to Rust style and keep maintaining alignment with
cdi-go code for specs-go.

Fixes #5
